backend/requirements.txt
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
python-magic==0.4.27
Faker==19.6.2
drf-spectacular==0.27.2
elasticsearch==8.12.0
certifi
